+++ controllers/about-controller.js
@@ -1,7 +1,10 @@
+import { accountsController } from "./accounts-controller.js";
+
 export const aboutController = {
-  index(request, response) {
+  async index(request, response) {
     const viewData = {
       title: "About Playlist",
+      user: await accountsController.getLoggedInUser(request)
     };
     console.log("about rendering");
     response.render("about-view", viewData);