+++ controllers/dashboard-controller.js
@@ -1,10 +1,12 @@
 import { stationStore } from "../models/station-store.js";
+import { accountsController } from "./accounts-controller.js";
 
 export const dashboardController = {
   async index(request, response) {
     const viewData = {
       title: "Stations Dashboard",
       stations: await stationStore.getAllStations(),
+      user: await accountsController.getLoggedInUser(request)
     };
     console.log("dashboard rendering");
     response.render("dashboard-view", viewData);
@@ -13,6 +15,9 @@ export const dashboardController = {
   async addStation(request, response) {
     const newStation = {
       title: request.body.title,
+      latitude: request.body.latitude,
+      longitude: request.body.longitude,
+      userId: request.body.userId,
     };
     console.log(`adding a New Station ${newStation.title}`);
     await stationStore.addStation(newStation);